+++ odoo/local-src/mtsmte_project/models/__init__.py
@@ -3,3 +3,5 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html)
 
 from . import project_project
+from . import sale_order
+from . import account_invoice